move times objects and change alerts file to use role
Moved the times objects from the alarm clock function to an export module in the alertTimesObj file. Changed the alerts export to be a function that takes a role and returns a random alert string. Made the relevant changes to the index.js file for the altered exports above as well as creating a function to determine which role should be @mentioned.